Added 5 City level tilemaps with the shitty tiles
I am using the shitty tiles so that the artists give me what I want and that is the tiles for the buildings. If they don't I am going to make sure it stays like this until SGX. It doesn't have to look good it just has to be functional for me to get a good grade. Call this strong arm tactics or what you want I don't care, its effective is what it is.